backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dotenv==1.0.0
pymongo==4.6.0
dnspython==2.4.2